move final rf read and scan function to rf_scan file
