+++ src/modules/rf/record.h
@@ -4,7 +4,6 @@
 #include "core/display.h"
 #include "emit.h"
 #include "record.h"
-#include "rf.h"
 #include "save.h"
 #include "structs.h"
